cpu data formtted
